Improved mix discovery and tagging. Faster refreshes via threads.
In addition to fetching the most popular tracks overall, we now also fetch the most popular tracks by tag and genre. This provides much more depth and leads to better mix quality overall (as determined by scores).
To complete refreshes in a reasonable amount of time we now create a thread for each request, this speeds up the ~160 api calls considerably.